refactor: extract settings modal into modules/settings.js
Moves SECTION 9 (preferences and the settings modal) out of app.js, which
drops it from 12,879 to 12,337 lines.

Canvas background and terminal font are now owned by the module. The other
preferences still live in app.js, so the module reaches them through the
injected context rather than importing them: an ES module cannot assign to
an imported binding, and these are all reassigned when the user changes a
setting. This matches the initGitGraphDeps/initNotificationDeps pattern the
other modules already use.

Two deliberate changes while moving the code:
- buildPickerItem was declared but never called, and is dropped.
- The ✓ and ▶ escapes are written as the literal characters, as
  the file is UTF-8.

The modal markup is byte-identical to the original, and the handler bodies
differ only where a mutated variable became a ctx setter call.
